Fix UE5 rendering: correct FMinimalViewInfo layout and all engine offsets
Major fixes:
- FRotator now uses double precision (TRotator<double>) matching UE5 LWC
- FOV offset corrected from 0x18/0x24 to 0x30 within FMinimalViewInfo
- FVector already double (0x18), FRotator now also double (0x18), FOV at 0x30
- CameraCacheEntry POV size updated from 0x5e0 to 0x850

All UE5 offsets ported from previous research branch:
- GWorld: 0x72aa2e0 -> 0xCA3CD68
- GName: 0x7129580 -> 0xC788078
- GObjects: 0xC8B3960 (new)
- CameraManager: 0x2c0 -> 0x328
- CameraCachePrivate: 0x1af0 -> 0x1460
- AcknowledgedPawn: 0x2a8 -> 0x370
- PlayerState: 0x248 -> 0x2D0
- OwningGameInstance: 0x180 -> 0x1E0
- RootComponent: 0x138 -> 0x1C0
- RelativeLocation: 0x11c -> 0x128
- HealthOffset: 0x1df8 -> 0x26D0
- TeamID: 0x400 -> 0x500

Also includes class-pointer based actor detection, debug validation
output, and explicit double->float casts in Camera.cpp.

+++ SquadDMA/SDK/Camera.cpp
@@ -42,8 +42,8 @@ static ViewMatrix CreateMatrix(Vector3 rot, Vector3 origin) {
 Vector2 Camera::WorldToScreen(MinimalViewInfo viewinfo, Vector3 world)
 {
 	Vector3 Screenlocation(0, 0, 0);
-	Vector3 rot = Vector3(viewinfo.Rotation.Pitch, viewinfo.Rotation.Yaw, viewinfo.Rotation.Roll);
-	Vector3 campos = Vector3(viewinfo.Location.X, viewinfo.Location.Y, viewinfo.Location.Z);
+	Vector3 rot = Vector3(static_cast<float>(viewinfo.Rotation.Pitch), static_cast<float>(viewinfo.Rotation.Yaw), static_cast<float>(viewinfo.Rotation.Roll));
+	Vector3 campos = Vector3(static_cast<float>(viewinfo.Location.X), static_cast<float>(viewinfo.Location.Y), static_cast<float>(viewinfo.Location.Z));
 	const ViewMatrix tempMatrix = CreateMatrix(rot, Vector3(0, 0, 0));
 
 	Vector3 vAxisX(tempMatrix.matrix[0][0], tempMatrix.matrix[0][1], tempMatrix.matrix[0][2]);
